backend/requirements.txt
python-dotenv==1.0.0
bcrypt==4.1.2
pyswip==0.2.10
flask==3.0.0
flask-restful==0.3.10
flask-compress==1.14.0
peewee==3.17.0
pdf2image==1.16.3
python-dateutil